backend/main.py
from fastapi import FastAPI, File, UploadFile
from fastapi.middleware.cors import CORSMiddleware
from ultralytics import YOLO
from PIL import Image
import io
import uvicorn
import torch
from torchvision import models, transforms
import torch.nn.functional as F
import numpy as np
from pathlib import Path

app = FastAPI()

# Enable CORS for frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"], 
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# --- LOAD MODELS ---
# 1. YOLOv11 (Detection)
YOLO_WEIGHTS = "runs/detect/backend/models/yolo_plant_detection18/weights/best.pt"
try:
    if Path(YOLO_WEIGHTS).exists():
        yolo_model = YOLO(YOLO_WEIGHTS)
        print(f"Trained YOLOv11 loaded from {YOLO_WEIGHTS}.")
    else:
        yolo_model = YOLO("yolo11n.pt")
        print("Trained YOLO weights not found. Loaded base yolo11n.pt.")
except Exception as e:
    print(f"Error loading YOLO: {e}")
    yolo_model = None

# 2. Disease Classifier (EfficientNet-B0)
CLASSIFIER_WEIGHTS = "backend/models/classifier_efficientnet_b0_best.pth"
try:
    if Path(CLASSIFIER_WEIGHTS).exists():
        # Initialize EfficientNet architecture
        classifier = models.efficientnet_b0(weights=None)
        num_features = classifier.classifier[1].in_features
        classifier.classifier[1] = torch.nn.Linear(num_features, 38)
        
        # Load custom weights
        checkpoint = torch.load(CLASSIFIER_WEIGHTS, map_location=torch.device('cpu'))
        classifier.load_state_dict(checkpoint['model_state_dict'])
        classifier.eval()
        print(f"Trained Disease Classifier (EfficientNet-B0) loaded from {CLASSIFIER_WEIGHTS}.")
    else:
        # Fallback to pre-trained ResNet if custom weights aren't available
        classifier = models.resnet50(weights=models.ResNet50_Weights.IMAGENET1K_V1)
        classifier.eval()
        print("Trained classifier weights not found. Loaded pre-trained ResNet50.")
except Exception as e:
    print(f"Error loading Classifier: {e}")
    classifier = None

# Image Preprocessing for Classifier
preprocess = transforms.Compose([
    transforms.Resize(256),
    transforms.CenterCrop(224),
    transforms.ToTensor(),
    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
])

# --- REAL DISEASE KNOWLEDGE BASE (38 Classes) ---
PLANT_CLASSES = [
    "Apple___Apple_scab", "Apple___Black_rot", "Apple___Cedar_apple_rust", "Apple___healthy",
    "Blueberry___healthy", "Cherry_(including_sour)___Powdery_mildew", "Cherry_(including_sour)___healthy",
    "Corn_(maize)___Cercospora_leaf_spot Gray_leaf_spot", "Corn_(maize)___Common_rust_",
    "Corn_(maize)___Northern_Leaf_Blight", "Corn_(maize)___healthy", "Grape___Black_rot",
    "Grape___Esca_(Black_Measles)", "Grape___Leaf_blight_(Isariopsis_Leaf_Spot)", "Grape___healthy",
    "Orange___Haunglongbing_(Citrus_greening)", "Peach___Bacterial_spot", "Peach___healthy",
    "Pepper,_bell___Bacterial_spot", "Pepper,_bell___healthy", "Potato___Early_blight",
    "Potato___Late_blight", "Potato___healthy", "Raspberry___healthy", "Soybean___healthy",
    "Squash___Powdery_mildew", "Strawberry___Leaf_scorch", "Strawberry___healthy",
    "Tomato___Bacterial_spot", "Tomato___Early_blight", "Tomato___Late_blight", "Tomato___Leaf_Mold",
    "Tomato___Septoria_leaf_spot", "Tomato___Spider_mites Two-spotted_spider_mite",
    "Tomato___Target_Spot", "Tomato___Tomato_Yellow_Leaf_Curl_Virus", "Tomato___Tomato_mosaic_virus",
    "Tomato___healthy"
]

def get_disease_details(class_idx, confidence):
    class_name = PLANT_CLASSES[class_idx]
    
    # Extract crop and disease from name
    parts = class_name.split('___')
    crop = parts[0].replace('_', ' ')
    disease = parts[1].replace('_', ' ')
    
    is_healthy = "healthy" in disease.lower()
    
    if is_healthy:
        return {
            "disease": f"{crop} (Healthy)",
            "description": f"The {crop} plant appears healthy with no visible signs of infection.",
            "risk": "Low",
            "loss": 0,
            "recs": ["Maintain regular watering", "Monitor for pests weekly"]
        }
    
    # Generic risk/loss mapping based on disease keywords
    risk = "Medium"
    loss = 15
    recs = ["Apply targeted fungicide/bactericide", "Remove infected foliage", "Improve air circulation"]
    
    if any(k in disease.lower() for k in ["blight", "virus", "rust", "curl"]):
        risk = "High"
        loss = 30
        recs.append("Isolate infected area to prevent spread")
    elif any(k in disease.lower() for k in ["spot", "mildew", "scab", "rot"]):
        risk = "Medium"
        loss = 15
        
    return {
        "disease": f"{crop} - {disease}",
        "description": f"Detected symptoms characteristic of {disease} in the {crop} specimen.",
        "risk": risk,
        "loss": loss,
        "recs": list(set(recs)),
        "crop": crop,
        "raw_disease": disease
    }


@app.get("/")
def read_root():
    return {"status": "AgroGuard Offline AI Server Running", "yolo": yolo_model is not None, "classifier": classifier is not None}

@app.post("/analyze")
async def analyze_crop(file: UploadFile = File(...)):
    if not yolo_model or not classifier:
        return {"error": "Models not loaded"}

    contents = await file.read()
    image = Image.open(io.BytesIO(contents)).convert('RGB')
    
    # 1. YOLO Detection
    yolo_results = yolo_model(image)
    plant_detected = False
    
    # Check for plants (using COCO classes: 46-banana, 47-apple, 49-orange, 50-broccoli, 51-carrot, 58-potted plant)
    # Broaden plant class check (COCO items related to plants/nature)
    plant_classes = [46, 47, 49, 50, 51, 58, 62, 63, 64] # banana, apple, orange, broccoli, carrot, potted plant, etc.
    for r in yolo_results:
        for c in r.boxes.cls:
            if int(c) in plant_classes or float(r.boxes.conf[0]) > 0.05: # High sensitivity
                plant_detected = True
                break
    
    # If YOLO implies it's definitely NOT a plant (and we trust it), we could return early.
    # However, for user experience, we might want to try classifying anyway or trigger a "No Plant" warning.
    # For this implementation, we will proceed but flag it.
    
    # 2. Disease Classification (CNN)
    input_tensor = preprocess(image)
    input_batch = input_tensor.unsqueeze(0)

    with torch.no_grad():
        output = classifier(input_batch)
    
    probabilities = F.softmax(output[0], dim=0)
    
    # Get Top 3 Differential Diagnoses
    top_prob, top_idx = torch.topk(probabilities, 3)
    main_diagnosis = get_disease_details(int(top_idx[0]), float(top_prob[0]))
    
    alternative_diagnoses = []
    for i in range(1, 3):
        alt_data = get_disease_details(int(top_idx[i]), float(top_prob[i]))
        alternative_diagnoses.append({
            "disease": alt_data["disease"],
            "confidence": round(float(top_prob[i]), 2)
        })

    # 3. Multi-Leaf Localized Analysis
    leaf_detections = []
    unique_diseases_found = set([main_diagnosis["disease"]])
    
    # COCO classes or our custom classes (1: healthy_leaf, 2: diseased_leaf)
    leaf_classes = [1, 2] 
    
    for r in yolo_results:
        for i, box in enumerate(r.boxes):
            cls_id = int(box.cls[0])
            conf = float(box.conf[0])
            
            if cls_id in leaf_classes and conf > 0.25:
                # Get crop coordinates
                b = box.xyxy[0].cpu().numpy().astype(int)
                leaf_crop = image.crop((b[0], b[1], b[2], b[3]))
                
                # Classify leaf crop
                crop_tensor = preprocess(leaf_crop).unsqueeze(0)
                with torch.no_grad():
                    crop_output = classifier(crop_tensor)
                crop_probs = F.softmax(crop_output[0], dim=0)
                crop_conf, crop_idx = torch.max(crop_probs, 0)
                
                leaf_diag = get_disease_details(int(crop_idx), float(crop_conf))
                unique_diseases_found.add(leaf_diag["disease"])
                
                leaf_detections.append({
                    "id": f"leaf_{i}",
                    "label": "Diseased Leaf" if cls_id == 2 else "Healthy Leaf",
                    "confidence": round(conf, 2),
                    "disease": leaf_diag["disease"],
                    "bbox": b.tolist()
                })

    # Update description if multiple conditions found
    if len(unique_diseases_found) > 1:
        extra_diseases = [d for d in unique_diseases_found if d != main_diagnosis["disease"]]
        main_diagnosis["description"] += f" Also identified: {', '.join(extra_diseases[:3])}."

    # If YOLO didn't see a plant, significantly lower confidence
    final_confidence = float(top_prob[0])
    if not plant_detected:
        main_diagnosis["description"] = f"NOTE: Detection uncertain. {main_diagnosis['description']}"
        final_confidence = final_confidence * 0.5 

    # Construct Response matching the Frontend 'YieldAnalysis' Interface
    response = {
        "expectedLoss": main_diagnosis["loss"],
        "confidenceScore": round(final_confidence, 2),
        "riskLevel": main_diagnosis["risk"],
        "recommendations": main_diagnosis["recs"],
        "diseaseDetected": main_diagnosis["disease"],
        "diseaseDescription": main_diagnosis["description"],
        "similarityScore": round(final_confidence, 2),
        "symptomlessStressDetected": "healthy" not in main_diagnosis["disease"].lower() and main_diagnosis["loss"] < 10,
        "stressProbability": main_diagnosis["loss"] * 2,
        "treatmentUrgency": "Immediate" if main_diagnosis["risk"] == "High" else ("Within 48h" if main_diagnosis["risk"] == "Medium" else "Monitoring"),
        "detailedMetrics": {
            "leafCoverage": main_diagnosis["loss"] * 1.5,
            "spreadVelocity": "Aggressive" if main_diagnosis["risk"] == "High" else "Moderate",
            "climateRiskFactor": 0.4
        },
        "alternativeDiagnoses": alternative_diagnoses,
        "leafDetections": leaf_detections
    }
    
    return response

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
