backend/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
torch
torchvision
tqdm
pyyaml
numpy
